@@ -1,14 +1,10 @@
 import { LightningElement } from 'lwc';
 import { getRawResults } from 'my/resultsFetcher';
-
+import type { GameResultsData } from 'my/resultsFetcher';
 
 export default class RawResultsTable extends LightningElement {
-
-    tableData = [];
+    tableData: GameResultsData[] = [];
     async connectedCallback() {
         this.tableData = await getRawResults();
     }
 }
-
-
-
