@@ -1,13 +1,13 @@
 interface Cookie {
     key: string;
-    value: string;
+    value: string | undefined;
 }
 
 function toCookie(str: string): Cookie {
     const eq = str.indexOf('=');
     if (eq >= 1) {
         const key = str.substring(0, eq);
-        const value = str.substring(eq+1);
+        const value = str.substring(eq + 1);
         return {
             key,
             value
@@ -27,8 +27,8 @@ function getCookie(key: string): Cookie {
     return getCookies().filter((c) => c.key === key)[0];
 }
 
-export function getUsername(): string {
-    const c = getCookie('dominionstats-username')
+export function getUsername(): string | undefined {
+    const c = getCookie('dominionstats-username');
     if (c && typeof c.value === 'string') {
         return decodeURIComponent(c.value);
     } else {
