@@ -1,16 +1,10 @@
 import { LightningElement } from 'lwc';
 import { getRawResults, extractaGameSizePlayerStats } from 'my/resultsFetcher';
-
+import type { PlayerStatsAllGames } from 'my/resultsFetcher';
 
 export default class sixPlayerRankingTable extends LightningElement {
-
-    tableData = [];
+    tableData: PlayerStatsAllGames[] = [];
     async connectedCallback() {
-        this.tableData = extractaGameSizePlayerStats(await getRawResults(),6);
+        this.tableData = extractaGameSizePlayerStats(await getRawResults(), 6);
     }
 }
-
-
-
-
-
