@@ -1,3 +1,4 @@
+import { createServer } from 'lwr';
 import bodyParser from 'body-parser';
 import compression from 'compression';
 import pgSession from 'connect-pg-simple';
@@ -10,10 +11,18 @@ import passport from 'passport';
 import passportGoogle from 'passport-google-oauth20';
 
 import type { DominionUser } from './common';
-import { init, getPool, testQueryAll, getGameResultsFromDb, insertGameResults } from './db_setup';
+import {
+    init,
+    getPool,
+    testQueryAll,
+    getGameResultsFromDb,
+    insertGameResults
+} from './db_setup';
 
 function setupRoutes() {
-    const app = express();
+    const lwrServer = createServer({ serverType: 'express' });
+    const app = lwrServer.getInternalServer<'express'>();
+
     app.use(compression());
 
     app.use(bodyParser.urlencoded({ extended: true }));
@@ -27,31 +36,33 @@ function setupRoutes() {
         SCHEME = 'https://';
         PUBLIC_PORT = '443';
         app.use(helmet());
-        app.use(helmet({
-            contentSecurityPolicy: {
-                useDefaults: true,
-                directives: {
-                    "upgrade-insecure-requests": null
-            
-                }
-            },
-            noSniff: undefined,
-        }));
+        app.use(
+            helmet({
+                contentSecurityPolicy: {
+                    useDefaults: true,
+                    directives: {
+                        'upgrade-insecure-requests': null
+                    }
+                },
+                noSniff: undefined
+            })
+        );
     }
 
     const GOOGLE_CLIENT_ID = process.env.GOOGLE_CLIENT_ID;
     const GOOGLE_CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET;
     if (!GOOGLE_CLIENT_ID || !GOOGLE_CLIENT_SECRET) {
-        throw new Error("Missing environment variables: GOOGLE_CLIENT_ID or GOOGLE_CLIENT_SECRET");
+        throw new Error(
+            'Missing environment variables: GOOGLE_CLIENT_ID or GOOGLE_CLIENT_SECRET'
+        );
     }
     const SESSION_SECRET = process.env.SESSION_SECRET;
     if (!SESSION_SECRET) {
-        throw new Error("Missing environment variable: SESSION_SECRET");
+        throw new Error('Missing environment variable: SESSION_SECRET');
     }
-    const DIST_DIR = './dist';
     const ALLOWLIST_EMAILS: string[] = (process.env.ALLOWLIST || '').split(',');
 
-    passport.serializeUser(function(user, done) {
+    passport.serializeUser(function (user, done) {
         done(null, user);
     });
 
@@ -72,7 +83,7 @@ function setupRoutes() {
         return true;
     }
 
-    passport.deserializeUser(function(user, done) {
+    passport.deserializeUser(function (user, done) {
         try {
             if (assertIsUser(user)) {
                 return done(null, user);
@@ -83,7 +94,7 @@ function setupRoutes() {
         }
     });
 
-    function ensureLoggedIn(options: {throw?: boolean}) {
+    function ensureLoggedIn(options: { throw?: boolean }) {
         // eslint-disable-next-line no-undef
         return (req: Express.Request, res: any, next: NextFunction) => {
             if (!req.isAuthenticated || !req.isAuthenticated()) {
@@ -94,46 +105,83 @@ function setupRoutes() {
             }
             next();
             return undefined;
-        }
+        };
     }
-    passport.use(new passportGoogle.Strategy({
-        clientID: GOOGLE_CLIENT_ID,
-        clientSecret: GOOGLE_CLIENT_SECRET,
-        callbackURL: `${SCHEME}${HOST}:${PUBLIC_PORT}/oauth2/redirect/accounts.google.com`,
-        scope: [ 'profile', 'email' ]
-    }, function (accessToken: string, refreshToken: string, profile: passportGoogle.Profile, cb: passportGoogle.VerifyCallback) {
-        if (!profile || !profile.emails || profile.emails.length === 0 || !profile.emails[0] || !profile.emails[0].value || !profile.emails[0].verified) {
-            console.error("User has no email attached or is unverified: ", profile);
-            return cb("User has no email attached or is unverified: " + profile?.emails);
-        }
-        const emails = profile?.emails?.filter((ev) => ev.verified);
-        if (!emails) {
-            return cb("User has no verified emails");
-        }
+    passport.use(
+        new passportGoogle.Strategy(
+            {
+                clientID: GOOGLE_CLIENT_ID,
+                clientSecret: GOOGLE_CLIENT_SECRET,
+                callbackURL: `${SCHEME}${HOST}:${PUBLIC_PORT}/oauth2/redirect/accounts.google.com`,
+                scope: ['profile', 'email']
+            },
+            function (
+                accessToken: string,
+                refreshToken: string,
+                profile: passportGoogle.Profile,
+                cb: passportGoogle.VerifyCallback
+            ) {
+                if (
+                    !profile ||
+                    !profile.emails ||
+                    profile.emails.length === 0 ||
+                    !profile.emails[0] ||
+                    !profile.emails[0].value ||
+                    !profile.emails[0].verified
+                ) {
+                    console.error(
+                        'User has no email attached or is unverified: ',
+                        profile
+                    );
+                    return cb(
+                        'User has no email attached or is unverified: ' +
+                            profile?.emails
+                    );
+                }
+                const emails = profile?.emails?.filter((ev) => ev.verified);
+                if (!emails) {
+                    return cb('User has no verified emails');
+                }
 
-        const allowedEmail = emails.find((ev) => ALLOWLIST_EMAILS.includes(ev.value));
-        if (!allowedEmail) {
-            return cb("User is not authorized to use these features");
-        }
+                const allowedEmail = emails.find((ev) =>
+                    ALLOWLIST_EMAILS.includes(ev.value)
+                );
+                if (!allowedEmail) {
+                    return cb('User is not authorized to use these features');
+                }
 
-        const user: DominionUser = {email: allowedEmail.value, name: profile.displayName };
-        return cb(null, user)
-    }));
+                const user: DominionUser = {
+                    email: allowedEmail.value,
+                    name: profile.displayName
+                };
+                return cb(null, user);
+            }
+        )
+    );
 
     app.use(cookieParser());
     if (!process.env.NODB && !process.env.INMEM_SESSION) {
-        app.use(expressSession({
-            store: new (pgSession(expressSession))({
-            // Insert connect-pg-simple options here
-            pool: getPool()
-            }),
-            secret: SESSION_SECRET,
-            resave: false,
-            cookie: { maxAge: 24 * 60 * 60 * 1000 } // 1 days
-            // Insert express-session options here
-        }));
+        app.use(
+            expressSession({
+                store: new (pgSession(expressSession))({
+                    // Insert connect-pg-simple options here
+                    pool: getPool()
+                }),
+                secret: SESSION_SECRET,
+                resave: false,
+                cookie: { maxAge: 24 * 60 * 60 * 1000 } // 1 days
+                // Insert express-session options here
+            })
+        );
     } else {
-        app.use(expressSession({secret: SESSION_SECRET, resave: false, saveUninitialized: false, cookie: { maxAge: 24*3600*1000 }}));
+        app.use(
+            expressSession({
+                secret: SESSION_SECRET,
+                resave: false,
+                saveUninitialized: false,
+                cookie: { maxAge: 24 * 3600 * 1000 }
+            })
+        );
     }
     app.use(passport.initialize());
     app.use(passport.session());
@@ -144,11 +192,15 @@ function setupRoutes() {
         req.logout();
         res.clearCookie(COOKIE_USERNAME);
         res.redirect('/');
-    })
+    });
 
-    app.get('/oauth2/redirect/accounts.google.com',
-        passport.authenticate('google', { failureRedirect: '/', failureMessage: true }),
-        function(req, res, next) {
+    app.get(
+        '/oauth2/redirect/accounts.google.com',
+        passport.authenticate('google', {
+            failureRedirect: '/',
+            failureMessage: true
+        }),
+        function (req, res, next) {
             if (req.user && assertIsUser(req.user)) {
                 if (req.user?.name !== req.cookies[COOKIE_USERNAME]) {
                     // if user successfully signed in, store user.name in cookie
@@ -156,7 +208,7 @@ function setupRoutes() {
                         res.cookie(COOKIE_USERNAME, req.user.name, {
                             // expire in year 9999 (from: https://stackoverflow.com/a/28289961)
                             expires: new Date(253402300000000),
-                            httpOnly: false, // allows JS code to access it
+                            httpOnly: false // allows JS code to access it
                         });
                     } else {
                         res.clearCookie(COOKIE_USERNAME);
@@ -165,28 +217,36 @@ function setupRoutes() {
             }
             next();
         },
-        function(req, res, next) {
+        function (req, res) {
             res.redirect('/');
-            next();
-    });
+        }
+    );
     app.get('/guarded', ensureLoggedIn({}), (req, res) => {
         res.send('Logged in as: ' + JSON.stringify(req.user));
     });
 
-    app.get('/api/v1/testObjects', ensureLoggedIn({throw: true}), async (req, res) => {
-        if (process.env.NODB) {
-            return res.status(501).send();
+    app.get(
+        '/api/v1/testObjects',
+        ensureLoggedIn({ throw: true }),
+        async (req, res) => {
+            if (process.env.NODB) {
+                return res.status(501).send();
+            }
+            return res.json(await testQueryAll());
         }
-        return res.json(await testQueryAll());
-    });
+    );
 
-    app.post('/api/v1/gameResults', ensureLoggedIn({throw: true}), async (req, res) => {
-        if (process.env.NODB) {
-            return res.status(501).send();
+    app.post(
+        '/api/v1/gameResults',
+        ensureLoggedIn({ throw: true }),
+        async (req, res) => {
+            if (process.env.NODB) {
+                return res.status(501).send();
+            }
+            const insertResult = await insertGameResults(req.body);
+            return res.status(insertResult.status).json(insertResult.results);
         }
-        const insertResult = await insertGameResults(req.body);
-        return res.status(insertResult.status).json(insertResult.results);
-    });
+    );
 
     app.get('/api/v1/gameResults', async (req, res) => {
         if (process.env.NODB) {
@@ -195,24 +255,37 @@ function setupRoutes() {
         return res.json(await getGameResultsFromDb());
     });
 
+    /*
     // Serve LWC content
     app.use(express.static(DIST_DIR));
 
     app.use('/', (req, res) => {
         res.sendFile(path.resolve(DIST_DIR, 'index.html'));
     });
+    */
 
-    app.listen(PORT, () =>
-        console.log(`✅  Server started: http://${HOST}:${PORT}`)
-    );
+    lwrServer
+        .listen(
+            ({ port, serverMode }: { port: Number; serverMode: string }) => {
+                console.log(
+                    `✅  ${serverMode} Server started: http://${HOST}:${port}`
+                );
+            }
+        )
+        .catch((err: Error) => {
+            console.error(err);
+            process.exit(1);
+        });
 }
 
 if (!process.env.NODB) {
     // Verify connection and run migrations on startup
-    init().catch((e) => {
-        console.error("Failed to init db_setup: ", e);
-        process.exit(1);
-    }).then(setupRoutes);
+    init()
+        .catch((e) => {
+            console.error('Failed to init db_setup: ', e);
+            process.exit(1);
+        })
+        .then(setupRoutes);
 } else {
     setupRoutes();
 }
