@@ -4,7 +4,10 @@ const invalidChars = /[!@#$%^&*()+=[\]{};':"\\|,.<>/?]+/; //used to make sure no
 const invalidPrefix = /^[ \t/]+.*/;
 const invalidSuffix = /.*[ \t]+$/;
 
-export function validateGameData(gameId: string, gameResults: PlayerResultForm[]): ErrorObject[] {
+export function validateGameData(
+    gameId: string,
+    gameResults: PlayerResultForm[]
+): ErrorObject[] {
     const errors: ErrorObject[] = [];
 
     if (!gameResults) {
@@ -20,14 +23,14 @@ export function validateGameData(gameId: string, gameResults: PlayerResultForm[]
     if (gameResults.length < 2) {
         errors.push({
             status: 'error',
-            error: 'Must enter a minimum of 2 players',
+            error: 'Must enter a minimum of 2 players'
         });
     }
 
     if (gameResults.length > 6) {
         errors.push({
             status: 'error',
-            error: "More than 6 players entered / this shouldn't be possible given there are only 6 places to input names",
+            error: "More than 6 players entered / this shouldn't be possible given there are only 6 places to input names"
         });
     }
 
@@ -39,27 +42,30 @@ export function validateGameData(gameId: string, gameResults: PlayerResultForm[]
         });
     }
 
-    gameResults.forEach(({playerName, victoryPoints}) => {
-
+    gameResults.forEach(({ playerName, victoryPoints }) => {
         //server validation
-        if (playerName === null || playerName === "") {
+        if (playerName === null || playerName === '') {
             errors.push({
                 status: 'error',
-                error: `Invalid Player Name; cannot be null or empty-string`,
+                error: `Invalid Player Name; cannot be null or empty-string`
             });
-        } 
+        }
         if (victoryPoints === null || !Number.isInteger(victoryPoints)) {
             errors.push({
                 status: 'error',
-                error: `Invalid Victory Points for player ${playerName}: ${victoryPoints}`,
+                error: `Invalid Victory Points for player ${playerName}: ${victoryPoints}`
             });
         }
 
         //check to see if there are any invalid characters in the string (see above for invalid chars)
-        if (invalidChars.test(playerName) || invalidPrefix.test(playerName) || invalidSuffix.test(playerName)){ 
+        if (
+            invalidChars.test(playerName) ||
+            invalidPrefix.test(playerName) ||
+            invalidSuffix.test(playerName)
+        ) {
             errors.push({
                 status: 'error',
-                error: `Invalid characters in playerName: "${playerName}"`,
+                error: `Invalid characters in playerName: "${playerName}"`
             });
         }
     });
